@remotion/studio: Address timeline keyframe review feedback
